+++ Publisher/src/main/java/com/telefonia/repository/CoberturaMunicipioRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import com.telefonia.Entities.CoberturaMunicipio;
 @Repository
-public interface CoberturaMunicipioRepository extends CrudRepository<CoberturaMunicipio, Short>{
+public interface CoberturaMunicipioRepository extends CrudRepository<CoberturaMunicipio, Integer>{
     
 
 }
